requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
Django==4.1.6
djangorestframework==3.14.0
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
mplfinance==0.12.9b7
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
six==1.16.0
sqlparse==0.4.3
tenacity==8.2.2
tzdata==2023.3
websocket-client==1.5.1
